led_control/requirements.txt
hidapi>=0.14.0
# tkinter is part of Python's standard library, no need to specify it in requirements.txt 